backend/requirements.txt
flask
flask-cors
requests
python-dotenv
beautifulsoup4
lxml
pymongo
